requirements.txt
django
djangorestframework
pygments
django_extensions